Move SPDM tests out of emulator
So that they can be run on the FPGA.

I tested that the tests start successfully on the FPGA but fail partway
through due to PEC errors. I suspect this is related to a hardware I3C
buffering issue, but I'll continue investigating.
