+++ common/testing/src/spdm_responder_validator/doe.rs
@@ -1,13 +1,13 @@
 // Licensed under the Apache-2.0 license
 
-use crate::tests::doe_util::common::DoeUtil;
-use crate::tests::spdm_responder_validator::common::{
+use crate::doe_util::common::DoeUtil;
+use crate::spdm_responder_validator::common::{
     execute_spdm_attestation, execute_spdm_responder_validator, execute_spdm_tee_io_validator,
     SpdmValidatorRunner, SERVER_LISTENING,
 };
-use crate::tests::spdm_responder_validator::transport::{Transport, SOCKET_TRANSPORT_TYPE_PCI_DOE};
-use crate::tests::spdm_responder_validator::SpdmTestType;
-use mcu_testing_common::{sleep_emulator_ticks, wait_for_runtime_start, MCU_RUNNING};
+use crate::spdm_responder_validator::transport::{Transport, SOCKET_TRANSPORT_TYPE_PCI_DOE};
+use crate::spdm_responder_validator::SpdmTestType;
+use crate::{sleep_emulator_ticks, wait_for_runtime_start, MCU_RUNNING};
 use std::net::TcpListener;
 use std::process::exit;
 use std::sync::atomic::Ordering;
