+++ common/testing/src/spdm_responder_validator/mod.rs
@@ -10,3 +10,8 @@ pub enum SpdmTestType {
     SpdmTeeIoValidator,
     SpdmAttestation,
 }
+
+pub use common::{
+    execute_spdm_attestation, execute_spdm_responder_validator, SpdmValidatorRunner,
+    SERVER_LISTENING,
+};
