+++ emulator/app/Cargo.toml
@@ -104,12 +104,6 @@ test-mbox-sram = []
 test-mci = []
 test-mctp-ctrl-cmds = ["emulator-periph/test-mctp-ctrl-cmds"]
 test-mctp-user-loopback = ["emulator-periph/test-mctp-user-loopback"]
-test-mctp-spdm-responder-conformance = [
-    "emulator-periph/test-mctp-spdm-responder-conformance",
-]
-test-mctp-spdm-attestation = [
-    "emulator-periph/test-mctp-spdm-attestation",
-]
 test-mcu-mbox-driver = ["emulator-periph/test-mcu-mbox-driver"]
 test-mcu-mbox-soc-requester-loopback = []
 test-mcu-mbox-usermode = []
