+++ emulator/app/src/emulator.rs
@@ -16,7 +16,6 @@ use crate::dis;
 use crate::doe_mbox_fsm;
 use crate::elf;
 use crate::tests;
-use crate::tests::spdm_responder_validator::SpdmTestType;
 use caliptra_api_types::DeviceLifecycle;
 use caliptra_emu_bus::BusMmio;
 use caliptra_emu_bus::{Bus, Clock, Timer};
@@ -43,6 +42,7 @@ use mcu_testing_common::i3c_socket;
 use mcu_testing_common::i3c_socket_server::start_i3c_socket;
 use mcu_testing_common::mctp_transport::MctpTransport;
 use mcu_testing_common::mctp_util::base_protocol::LOCAL_TEST_ENDPOINT_EID;
+use mcu_testing_common::spdm_responder_validator::SpdmTestType;
 use mcu_testing_common::{MCU_RUNNING, MCU_RUNTIME_STARTED, MCU_TICKS, TICK_COND};
 use pldm_fw_pkg::FirmwareManifest;
 use pldm_ua::daemon::PldmDaemon;
@@ -651,37 +651,13 @@ impl Emulator {
                 spdm_loopback_tests,
                 None,
             );
-        } else if cfg!(feature = "test-mctp-spdm-responder-conformance") {
-            if std::env::var("SPDM_VALIDATOR_DIR").is_err() {
-                println!("SPDM_VALIDATOR_DIR environment variable is not set. Skipping test");
-                exit(0);
-            }
-            i3c_controller_join_handle = Some(i3c_controller.start());
-            crate::tests::spdm_responder_validator::mctp::run_mctp_spdm_conformance_test(
-                cli.i3c_port.unwrap(),
-                i3c.get_dynamic_address().unwrap(),
-                SpdmTestType::SpdmResponderConformance,
-                std::time::Duration::from_secs(9000), // timeout in seconds
-            );
-        } else if cfg!(feature = "test-mctp-spdm-attestation") {
-            if std::env::var("SPDM_VALIDATOR_DIR").is_err() {
-                println!("SPDM_VALIDATOR_DIR environment variable is not set. Skipping test");
-                exit(0);
-            }
-            i3c_controller_join_handle = Some(i3c_controller.start());
-            crate::tests::spdm_responder_validator::mctp::run_mctp_spdm_attestation_test(
-                cli.i3c_port.unwrap(),
-                i3c.get_dynamic_address().unwrap(),
-                SpdmTestType::SpdmAttestation,
-                std::time::Duration::from_secs(9000), // timeout in seconds
-            );
         } else if cfg!(feature = "test-doe-spdm-responder-conformance") {
             if std::env::var("SPDM_VALIDATOR_DIR").is_err() {
                 println!("SPDM_VALIDATOR_DIR environment variable is not set. Skipping test");
                 exit(0);
             }
             let (test_rx, test_tx) = doe_mbox_fsm.start();
-            crate::tests::spdm_responder_validator::doe::run_doe_spdm_conformance_test(
+            mcu_testing_common::spdm_responder_validator::doe::run_doe_spdm_conformance_test(
                 test_tx,
                 test_rx,
                 SpdmTestType::SpdmResponderConformance,
@@ -693,7 +669,7 @@ impl Emulator {
                 exit(0);
             }
             let (test_rx, test_tx) = doe_mbox_fsm.start();
-            crate::tests::spdm_responder_validator::doe::run_doe_spdm_conformance_test(
+            mcu_testing_common::spdm_responder_validator::doe::run_doe_spdm_conformance_test(
                 test_tx,
                 test_rx,
                 SpdmTestType::SpdmTeeIoValidator,
