+++ emulator/app/src/tests/doe_discovery.rs
@@ -1,8 +1,8 @@
 // Licensed under the Apache-2.0 license
 
 use crate::doe_mbox_fsm::{DoeTestState, DoeTransportTest};
-use crate::tests::doe_util::common::DoeUtil;
-use crate::tests::doe_util::protocol::*;
+use mcu_testing_common::doe_util::common::DoeUtil;
+use mcu_testing_common::doe_util::protocol::*;
 use mcu_testing_common::{sleep_emulator_ticks, MCU_RUNNING};
 use std::sync::atomic::Ordering;
 use std::sync::mpsc::{Receiver, Sender};
