+++ emulator/app/src/tests/mod.rs
@@ -4,9 +4,7 @@ pub mod caliptra_util_host_validator;
 pub mod doe_discovery;
 pub mod doe_transport_loopback;
 pub mod doe_user_loopback;
-pub mod doe_util;
 pub mod emulator_mcu_mailbox_test;
 pub mod mctp_ctrl_cmd;
 pub mod mctp_user_loopback;
 pub mod pldm_request_response_test;
-pub mod spdm_responder_validator;
