+++ emulator/periph/Cargo.toml
@@ -68,7 +68,5 @@ test-pldm-discovery = []
 test-pldm-fw-update = []
 test-pldm-fw-update-e2e = []
 test-pldm-streaming-boot = []
-test-mctp-spdm-responder-conformance = []
-test-mctp-spdm-attestation = []
 test-doe-spdm-responder-conformance = []
 test-doe-spdm-tdisp-ide-validator = []
