+++ platforms/emulator/runtime/userspace/apps/example/Cargo.toml
@@ -38,6 +38,7 @@ embedded-alloc.workspace = true
 [features]
 default = []
 debug = []
+doe = []
 hw-2-1 = []
 test-caliptra-certs = []
 test-caliptra-crypto = []
