+++ platforms/emulator/runtime/userspace/apps/user/Cargo.toml
@@ -51,6 +51,7 @@ portable-atomic = { workspace = true, features = ["critical-section"] }
 default = []
 debug = []
 hw-2-1 = []
+doe = []
 test-caliptra-certs = []
 test-caliptra-crypto = []
 test-caliptra-mailbox = []
@@ -95,9 +96,8 @@ test-pldm-discovery = []
 test-pldm-fw-update = []
 test-pldm-fw-update-e2e = []
 test-pldm-streaming-boot = []
-test-mctp-spdm-responder-conformance = []
 test-mctp-spdm-attestation = []
-test-doe-spdm-responder-conformance = []
-test-doe-spdm-tdisp-ide-validator = []
+test-doe-spdm-responder-conformance = ["doe"]
+test-doe-spdm-tdisp-ide-validator = ["doe"]
 test-mcu-mbox-fips-periodic = ["mcu-mbox-lib/periodic-fips-self-test"]
 active-i3c1 = []
